rename buildd image to just "build"
The 'buildd' image took its name from debootstrap's --variant
buildd, but it's a bit misleading: It's a build environment,
such as a buildd has, but a buildd is more than that too, it's
the listening daemon, queuing, etc. -- and we could conceivably
have a real buildd image in the future. So rename this one now
to just 'build'.