Reduce size of build context (fixes #1)
Restructure the repository to reduce the build context size by
removing the unpacked debootstrap output from CWD when "docker
build" is invoked.

Since the Makefiles for the different flavours are virtually
identical, take the opportunity to drop down to just one Makefile,
with suite, platform etc. parameterized via variables.

Add a 'build.sh' to invoke make for the common targets.