Adjust Makefile to survive running un-privileged
If the user runs "make", the rule that runs debootstrap is considered
satisfied, and subsequent privileged re-runs will fail because the
chroot is empty.

Add a layer of indirection that means the debootstrap has to have
been run for the rule to be satisfied, by depending on the "etc"
subdirectory which we do not explicitly create.

Closes: #7.